- configure.ac refined - OC now compiles and runs on Solaris 10 using the Universal TUN/TAP driver   version 1.1 (http://vtun.sourceforge.net/tun/index.html) - ocathosts.[ch] added. It reads IPv6 addresses from /etc/hosts - /etc/hosts reverse lookup for I2P .b32 names added

+++ trunk/src/ocatv6conv.c
@@ -80,7 +80,7 @@ int oniontipv6(const char *onion, struct in6_addr *ip6)
    for (i = 0; i < 16; i++)
    {
       shl5((char*) ip6);
-      j = toupper(onion[i]);
+      j = toupper((int) onion[i]);
       if ((j < '2') || (j > 'Z'))
          return -1;
       if ((j = deBASE32_[j - '2']) == -1)
